requirements.txt
discord.py==1.7.3
Flask==2.0.1
requests==2.25.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
idna==2.8
urllib3==1.25.8
mysql-connector-python